Move enums with permissions and roles into separated package
