+++ src/main/java/com/md/scheduler/configuration/security/enums/AppUserRole.java
@@ -1,12 +1,12 @@
-package com.md.scheduler.configuration.security;
+package com.md.scheduler.configuration.security.enums;
 
 import org.springframework.security.core.authority.SimpleGrantedAuthority;
 
 import java.util.Set;
 import java.util.stream.Collectors;
 
-import static com.md.scheduler.configuration.security.AppUserPermission.USER_READ;
-import static com.md.scheduler.configuration.security.AppUserPermission.USER_WRITE;
+import static com.md.scheduler.configuration.security.enums.AppUserPermission.USER_READ;
+import static com.md.scheduler.configuration.security.enums.AppUserPermission.USER_WRITE;
 
 public enum AppUserRole {
